requirements.txt
Django==1.10
django-extensions==1.7.3
Pillow==3.3.1
